# Conflicts:
#	fdbserver/DataDistribution.actor.cpp

+++ flow/Trace.cpp
@@ -1039,7 +1039,7 @@ void TraceBatch::dump() {
 	for(int i = 0; i < attachBatch.size(); i++) {
 		char buffer[256];
 		int length = sprintf(buffer, "<Event Severity=\"%d\" Time=\"%.6f\" Type=\"%s\" Machine=\"%d.%d.%d.%d:%d\" logGroup=\"%.*s\" ID=\"%016" PRIx64 "\" To=\"%016" PRIx64 "\"/>\r\n",
-			(int)SevInfo, attachBatch[i].time, attachBatch[i].name, (local.ip>>24)&0xff,(local.ip>>16)&0xff,(local.ip>>8)&0xff,local.ip&0xff,local.port, g_traceLog.logGroup.size(), g_traceLog.logGroup.data(),
+			(int)SevInfo, attachBatch[i].time, attachBatch[i].name, (local.ip>>24)&0xff,(local.ip>>16)&0xff,(local.ip>>8)&0xff,local.ip&0xff,local.port, (int)g_traceLog.logGroup.size(), g_traceLog.logGroup.data(),
 			attachBatch[i].id, attachBatch[i].to);
 		g_traceLog.write( buffer, length );
 	}
@@ -1047,7 +1047,7 @@ void TraceBatch::dump() {
 	for(int i = 0; i < eventBatch.size(); i++) {
 		char buffer[256];
 		int length = sprintf(buffer, "<Event Severity=\"%d\" Time=\"%.6f\" Type=\"%s\" Machine=\"%d.%d.%d.%d:%d\" logGroup=\"%.*s\" ID=\"%016" PRIx64 "\" Location=\"%s\"/>\r\n",
-			(int)SevInfo, eventBatch[i].time, eventBatch[i].name, (local.ip>>24)&0xff,(local.ip>>16)&0xff,(local.ip>>8)&0xff,local.ip&0xff,local.port, g_traceLog.logGroup.size(), g_traceLog.logGroup.data(),
+			(int)SevInfo, eventBatch[i].time, eventBatch[i].name, (local.ip>>24)&0xff,(local.ip>>16)&0xff,(local.ip>>8)&0xff,local.ip&0xff,local.port, (int)g_traceLog.logGroup.size(), g_traceLog.logGroup.data(),
 			eventBatch[i].id, eventBatch[i].location );
 		g_traceLog.write( buffer, length );
 	}
@@ -1055,7 +1055,7 @@ void TraceBatch::dump() {
 	for(int i = 0; i < buggifyBatch.size(); i++) {
 		char buffer[256];
 		int length = sprintf( buffer, "<Event Severity=\"%d\" Time=\"%.6f\" Type=\"BuggifySection\" Machine=\"%d.%d.%d.%d:%d\" logGroup=\"%.*s\" Activated=\"%d\" File=\"%s\" Line=\"%d\"/>\r\n",
-			(int)SevInfo, buggifyBatch[i].time, (local.ip>>24)&0xff,(local.ip>>16)&0xff,(local.ip>>8)&0xff,local.ip&0xff,local.port, g_traceLog.logGroup.size(), g_traceLog.logGroup.data(),
+			(int)SevInfo, buggifyBatch[i].time, (local.ip>>24)&0xff,(local.ip>>16)&0xff,(local.ip>>8)&0xff,local.ip&0xff,local.port, (int)g_traceLog.logGroup.size(), g_traceLog.logGroup.data(),
 			buggifyBatch[i].activated, buggifyBatch[i].file.c_str(), buggifyBatch[i].line );
 		g_traceLog.write( buffer, length );
 	}
